Think handle_command is finished
In order to test test1.in, I had to change the first line in the file from 3C031000 to 3C031001 in order for it to work correctly. This is because later in the program 3C031000 gives you a memory location of 0x10000000 which is actually reserved for $gp, and the memory for the program distinguished by the simulator actually starts at 0x10010000. So by changing to 3C031001, the memory location becomes 0x10010000. I emailed the TA and she said this was correct.